Move Form logic to their own namespace, refactoring the StoredForm & TempForm for a more robust structure

+++ tests/integration/SubmissionOverview/SubmissionOverviewTest.php
@@ -114,6 +114,8 @@ final class SubmissionOverviewTest extends TestCase {
 	}
 	
 	public static function tearDownAfterClass() {
-		static::$Form->delete();
+		if(!empty(static::$Form)) {
+			static::$Form->delete();
+		}
 	}
 }
